remove resource unused

+++ app/src/main/java/yovi/putra/hackernews/data/entity/Comment.kt
@@ -10,6 +10,6 @@ data class Comment(
     val kids: MutableList<String>?,
     val parent: Int?,
     val text: String?,
-    val time: Int?,
+    val time: Long?,
     val type: String?
 ) : Parcelable
